+++ src/distance_calculator.cpp
@@ -54,7 +54,7 @@ uint8_t DistanceCalculator::findNearestIntersection(float * origin)
 	uint8_t below = 0;
 	uint8_t above = 0;
 
-	for(size_t i = 0; i < mySTL_file->numberOfFacets; i++) {
+	for(uint32_t i = 0; i < mySTL_file->numberOfFacets; i++) {
 		if(mySTL_file->useFacet[i]) {
 			float vert0[3] = {
 				mySTL_file->surface[i].vertex1.x,
